Rename PriceForm -> PriceCalculator
