+++ apps/store/src/components/PriceCalculator/AutomaticField.tsx
@@ -1,8 +1,8 @@
 import { InputField } from 'ui'
 import { InputSelect } from '@/components/InputSelect/InputSelect'
-import { HouseholdSizeField } from '@/components/PriceForm/HouseholdSize'
-import { InputField as InputFieldType } from '@/services/PriceForm/Field.types'
-import { JSONData } from '@/services/PriceForm/PriceForm.types'
+import { HouseholdSizeField } from '@/components/PriceCalculator/HouseholdSize'
+import { InputField as InputFieldType } from '@/services/PriceCalculator/Field.types'
+import { JSONData } from '@/services/PriceCalculator/PriceCalculator.types'
 import { useProductPageContext } from '../ProductPage/ProductPageContext'
 import { CarRegistrationNumberField } from './CarRegistrationField'
 import { CurrentInsuranceField } from './CurrentInsuranceField/CurrentInsuranceField'
