+++ apps/store/src/components/PriceCalculator/InputSwitch.tsx
