+++ apps/store/src/components/PriceCalculator/TickIcon.tsx
