+++ apps/store/src/components/ProductPage/ProductPage.types.ts
@@ -1,5 +1,5 @@
 import { ProductDataQuery } from '@/services/apollo/generated'
-import { Template } from '@/services/PriceForm/PriceForm.types'
+import { Template } from '@/services/PriceCalculator/PriceCalculator.types'
 import { PriceIntent } from '@/services/priceIntent/priceIntent.types'
 import { ShopSession } from '@/services/shopSession/ShopSession.types'
 import { ProductStory, StoryblokPageProps } from '@/services/storyblok/storyblok'
