+++ apps/store/src/services/PriceCalculator/data/SE_HOUSE.ts
@@ -1,10 +1,7 @@
-import { Template } from '../PriceForm.types'
+import { Template } from '../PriceCalculator.types'
 
-export const NO_HOUSE: Template = {
-  name: 'NO_HOUSE',
-  initialData: {
-    isStudent: false,
-  },
+export const SE_HOUSE: Template = {
+  name: 'SE_HOUSE',
   sections: [
     {
       id: 'your-info',
@@ -13,9 +10,9 @@ export const NO_HOUSE: Template = {
       items: [
         {
           field: {
-            type: 'date',
-            name: 'birthDate',
-            label: { key: 'Date of birth' },
+            type: 'ssn-se',
+            name: 'ssn',
+            label: { key: 'Personal number' },
             required: true,
           },
           layout: { columnSpan: 6 },
@@ -40,9 +37,8 @@ export const NO_HOUSE: Template = {
           field: {
             type: 'text',
             name: 'zipCode',
+            inputMode: 'numeric',
             label: { key: 'Postal code' },
-            minLength: 4,
-            maxLength: 4,
             required: true,
           },
           layout: { columnSpan: 3 },
@@ -50,8 +46,18 @@ export const NO_HOUSE: Template = {
         {
           field: {
             type: 'number',
-            name: 'squareMeters',
-            label: { key: 'Square meters' },
+            name: 'livingSpace',
+            label: { key: 'House size' },
+            required: true,
+            min: 0,
+          },
+          layout: { columnSpan: 3 },
+        },
+        {
+          field: {
+            type: 'number',
+            name: 'ancillaryArea',
+            label: { key: 'Ancillary area' },
             required: true,
             min: 0,
           },
@@ -70,32 +76,10 @@ export const NO_HOUSE: Template = {
         {
           field: {
             type: 'number',
-            name: 'yearOfOwnership',
-            label: { key: 'Year of ownership' },
+            name: 'numberOfBathrooms',
+            label: { key: 'Number of bathrooms' },
             required: true,
             min: 0,
-          },
-          layout: { columnSpan: 3 },
-        },
-        {
-          field: {
-            type: 'number',
-            name: 'numberOfWetUnits',
-            label: { key: 'Number of wet units' },
-            required: true,
-            min: 0,
-          },
-          layout: { columnSpan: 3 },
-        },
-        {
-          field: {
-            type: 'select',
-            name: 'waterLeakageDetector',
-            label: { key: 'Water leakage detector' },
-            options: [
-              { label: { key: 'Yes' }, value: 'true' },
-              { label: { key: 'No' }, value: 'false' },
-            ],
           },
           layout: { columnSpan: 3 },
         },
@@ -104,6 +88,7 @@ export const NO_HOUSE: Template = {
             type: 'select',
             name: 'isSubleted',
             label: { key: 'Subletting' },
+            required: true,
             options: [
               { label: { key: 'Yes' }, value: 'true' },
               { label: { key: 'No' }, value: 'false' },
@@ -116,6 +101,7 @@ export const NO_HOUSE: Template = {
             type: 'extra-buildings',
             name: 'extraBuildings',
             label: { key: 'Extra buildings' },
+            defaultValue: [],
             buildingOptions: [
               { label: { key: 'Garage' }, value: 'GARAGE' },
               { label: { key: 'Carport' }, value: 'CARPORT' },
@@ -132,7 +118,6 @@ export const NO_HOUSE: Template = {
               { label: { key: 'Boathouse' }, value: 'BOATHOUSE' },
               { label: { key: 'Other' }, value: 'OTHER' },
             ],
-            defaultValue: [],
           },
           layout: { columnSpan: 6 },
         },
@@ -145,13 +130,12 @@ export const NO_HOUSE: Template = {
       items: [
         {
           field: {
-            type: 'number',
-            min: 0,
-            max: 5,
+            type: 'householdSize',
             name: 'numberCoInsured',
-            label: { key: 'Number of co-insured' },
+            label: { key: 'Household size' },
             required: true,
             defaultValue: 0,
+            max: 5,
           },
           layout: { columnSpan: 6 },
         },
